==> ASHRAE/11 Hyper parameter Tuning_LGBM.ipynb <==
# No leak data
import pandas as pd
import numpy as np
import gc 
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns
import datetime


# Bayesian Optimization Functions

# https://www.kaggle.com/aitude/ashrae-hyperparameter-tuning

train_df = pd.read_pickle('train_df.pkl')
test_df = pd.read_pickle('test_df.pkl')

# some feature enginnering

train_df['age'] = train_df['year_built'].max()-train_df['year_built']+1
test_df['age'] = test_df['year_built'].max() - test_df['year_built'] + 1

#Handling missing values
# To streamline this though process it is useful to know the 3 categories in which missing data can be classified into:

# Missing Completely at Random (MCAR)
# Missing at Random (MAR)
# Missing Not at Random (MNAR)

train_df['floor_count'] = train_df['floor_count'].fillna(-999).astype(np.int16)
test_df['floor_count'] = test_df['floor_count'].fillna(-999).astype(np.int16)

train_df['year_built'] = train_df['year_built'].fillna(-999).astype(np.int16)
test_df['year_built'] = test_df['year_built'].fillna(-999).astype(np.int16)

train_df['age'] = train_df['age'].fillna(-999).astype(np.int16)
test_df['age'] = test_df['age'].fillna(-999).astype(np.int16)

train_df['cloud_coverage'] = train_df['cloud_coverage'].fillna(-999).astype(np.int16)
test_df['cloud_coverage'] = test_df['cloud_coverage'].fillna(-999).astype(np.int16) 

train_df['square_feet'] = np.log1p(train_df['square_feet'])
test_df['square_feet'] = np.log1p(test_df['square_feet'])

def make_is_bad_zero(Xy_subset, min_interval=48, summer_start=3000, summer_end=7500):
    """Helper routine for 'find_bad_zeros'.
    
    This operates upon a single dataframe produced by 'groupby'. We expect an 
    additional column 'meter_id' which is a duplicate of 'meter' because groupby 
    eliminates the original one."""
    meter = Xy_subset.meter_id.iloc[0]
    is_zero = Xy_subset.meter_reading == 0
    if meter == 0:
        # Electrical meters should never be zero. Keep all zero-readings in this table so that
        # they will all be dropped in the train set.
        return is_zero

    transitions = (is_zero != is_zero.shift(1))
    all_sequence_ids = transitions.cumsum()
    ids = all_sequence_ids[is_zero].rename("ids")
    if meter in [2, 3]:
        # It's normal for steam and hotwater to be turned off during the summer
        keep = set(ids[(Xy_subset.timestamp < summer_start) |
                       (Xy_subset.timestamp > summer_end)].unique())
        is_bad = ids.isin(keep) & (ids.map(ids.value_counts()) >= min_interval)
    elif meter == 1:
        time_ids = ids.to_frame().join(Xy_subset.timestamp).set_index("timestamp").ids
        is_bad = ids.map(ids.value_counts()) >= min_interval

        # Cold water may be turned off during the winter
        jan_id = time_ids.get(0, False)
        dec_id = time_ids.get(8283, False)
        if (jan_id and dec_id and jan_id == time_ids.get(500, False) and
                dec_id == time_ids.get(8783, False)):
            is_bad = is_bad & (~(ids.isin(set([jan_id, dec_id]))))
    else:
        raise Exception(f"Unexpected meter type: {meter}")

    result = is_zero.copy()
    result.update(is_bad)
    return result

def find_bad_zeros(X, y):
    """Returns an Index object containing only the rows which should be deleted."""
    Xy = X.assign(meter_reading=y, meter_id=X.meter)
    is_bad_zero = Xy.groupby(["building_id", "meter"]).apply(make_is_bad_zero)
    return is_bad_zero[is_bad_zero].index.droplevel([0, 1])

def find_bad_sitezero(X):
    """Returns indices of bad rows from the early days of Site 0 (UCF)."""
    return X[(X.timestamp < 3378) & (X.site_id == 0) & (X.meter == 0)].index

def find_bad_building1099(X, y):
    """Returns indices of bad rows (with absurdly high readings) from building 1099."""
    return X[(X.building_id == 1099) & (X.meter == 2) & (y > 3e4)].index

def find_bad_rows(X, y):
    return find_bad_zeros(X, y).union(find_bad_sitezero(X)).union(find_bad_building1099(X, y))


# bad rows 삭제 3% 
train_df['timestamp'] = (train_df.timestamp - pd.to_datetime("2016-01-01")).dt.total_seconds() // 3600
rows = find_bad_rows(train_df.drop(['meter_reading'],axis =1 ),train_df['meter_reading'])
rows = list(rows)
train_df= train_df.drop(rows)

# drop_cols = ['date',"precip_depth_1_hr", "sea_level_pressure", "wind_direction", "wind_speed","timestamp"]
drop_cols = ['date',"timestamp"]
target = train_df["meter_reading_log1p"]
del train_df["meter_reading"], train_df['meter_reading_log1p']
train_df = train_df.drop(drop_cols, axis=1)
drop_cols += ["row_id"]
# drop_cols.remove('date')
test_df = test_df.drop(drop_cols, axis=1)

train_df.head()
categorical_features = ["building_id", "site_id", "meter", "hour", "weekend",'month','is_holiday','primary_use']

leak_df = pd.read_pickle('leak_df.pkl')
leak_df = leak_df.dropna()


%%time
import lightgbm as lgb
from sklearn.model_selection import KFold
from sklearn.metrics import mean_squared_log_error
from bayes_opt import BayesianOptimization
folds = 3
seed = 99 #666

# Hyperparameter range
params_range = {
                'num_leaves': (1000, 1500),
                'feature_fraction': (0.7, 0.9),
                'bagging_fraction': (0.8, 1),
                'max_depth': (10, 11),
                'lambda_l1': (0, 5),
                'lambda_l2': (0, 5),
                'min_split_gain': (0.001, 0.1),
                'min_child_weight': (5, 50),
                'learning_rate' : (.03,.07)
               }



%%time
paramlst = []
scorelst =[]
def running(num_leaves, feature_fraction, bagging_fraction, max_depth, lambda_l1, lambda_l2, min_split_gain, min_child_weight, learning_rate):
    params = {
    "objective": "regression",
    "boosting": "gbdt",
    "num_leaves": int(round(num_leaves)),
    "learning_rate": learning_rate,
    "feature_fraction": max(min(feature_fraction,1),0),
    "bagging_fraction": max(min(bagging_fraction,1),0),
    "max_depth" : int(round(max_depth)),
    "lambda_l1" : max(lambda_l1,0),
    "lambda_l2" : max(lambda_l2,0),
    "min_split_gain": min_split_gain,
    "min_child_weight": min_child_weight,
    "metric": "rmse",
    "n_jobs" : 10
    }
    
    print(params)
    kf = KFold(n_splits=folds, shuffle=False, random_state=seed)
    scores = [] 
    models = []
    for train_index, val_index in kf.split(train_df):
        train_X = train_df.iloc[train_index]
    
        val_X = train_df.iloc[val_index]
        train_y = target.iloc[train_index]
        val_y = target.iloc[val_index]
        lgb_train = lgb.Dataset(train_X, train_y,categorical_feature=categorical_features)
        lgb_eval = lgb.Dataset(val_X, val_y,categorical_feature=categorical_features)
        gbm = lgb.train(params,
                    lgb_train,
                    num_boost_round=2000, #300,
                    valid_sets=(lgb_train, lgb_eval),
#                     feval=rmsle,
                    early_stopping_rounds= 50,#100,
                    verbose_eval=30) #100)
        models.append(gbm)
        del gbm
    
    i = 0
    res3 = []
    step_size = 50000
        
    for j in tqdm(range(int(np.ceil(test_df.shape[0] / 50000)))):
        res3.append([np.expm1(model.predict(test_df.iloc[i:i + step_size])) for model in models])
        i += step_size
        
    res4 = []
    print(len(res3))
    for k in range(len(res3)):
        res4.append((0.6*res3[k][0]+0.3*res3[k][1]+0.1*res3[k][2]))
#         if i==100 : break
    
    
    res4 = np.concatenate(res4)
    sample_submission = pd.read_csv('sample_submission.csv')
    sample_submission["meter_reading"] = res4
    sample_submission.loc[sample_submission['meter_reading'] < 0, 'meter_reading'] = 0

    sample_submission = pd.merge(sample_submission,leak_df.loc[:,['row_id','leak_meter_reading']],how ='inner')    
    score = np.sqrt(mean_squared_log_error( sample_submission['leak_meter_reading'], sample_submission['meter_reading'] ))
    print(score)
    scorelst.append(score)
    paramlst.append(params)
    
    df_para = pd.DataFrame({"parameter" : paramlst,"score": scorelst})
    df_para.to_csv('para.csv',index=False)
    
    return -np.sqrt(mean_squared_log_error( sample_submission['leak_meter_reading'], sample_submission['meter_reading'] ))        

    
Baysian = BayesianOptimization(running, params_range,random_state = seed)
Baysian.maximize(init_points = 5, n_iter = 20)

# def lgb_best_params(X, y,opt_params,init_points=2, optimization_round=20, n_folds=3, random_seed=0, cv_estimators=1000):
    
#     # prepare dataset
#     categorical_features = ["building_id", "site_id", "meter", "primary_use", "is_holiday", "weekend"]
#     train_data = lgb.Dataset(data=X, label=y, categorical_feature = categorical_features, free_raw_data=False)
    
#     def lgb_run(num_leaves, feature_fraction, bagging_fraction, max_depth, lambda_l1, lambda_l2, min_split_gain, min_child_weight,learning_rate):
#         params = {"boosting": "gbdt",'application':'regression','num_iterations':cv_estimators, 'early_stopping_round':int(cv_estimators/5), 'metric':'rmse'}
#         params["num_leaves"] = int(round(num_leaves))
#         params['feature_fraction'] = max(min(feature_fraction, 1), 0)
#         params['bagging_fraction'] = max(min(bagging_fraction, 1), 0)
#         params['max_depth'] = int(round(max_depth))
#         params['lambda_l1'] = max(lambda_l1, 0)
#         params['lambda_l2'] = max(lambda_l2, 0)
#         params['min_split_gain'] = min_split_gain
#         params['min_child_weight'] = min_child_weight
#         params['learning_rate'] = learning_rate
#         cv_result = lgb.cv(params, train_data, nfold=n_folds, seed=random_seed, stratified=False, verbose_eval =cv_estimators, metrics=['rmse'])
#         return -min(cv_result['rmse-mean'])
    
#     params_finder = BayesianOptimization(lgb_run, opt_params, random_state=2021)
#     # optimize
#     params_finder.maximize(init_points=init_points, n_iter=optimization_round)

#     # return best parameters
#     return params_finder.max

# #  number of fold
# fold = 3

# # Hyperparameter range
# params_range = {
#                 'num_leaves': (1000, 1280),
#                 'feature_fraction': (0.7, 0.9),
#                 'bagging_fraction': (0.8, 1),
#                 'max_depth': (10, 11),
#                 'lambda_l1': (2, 5),
#                 'lambda_l2': (2, 5),
#                 'min_split_gain': (0.001, 0.1),
#                 'min_child_weight': (5, 50),
#                 'learning_rate' : (.05,.07)
#                }

# # You can experiments with different estimators in a single execution. I'm using small numbers for demonstration purpose so you must change to high numbers.
# cv_estimators = [50,100,200]

# #n_iter: How many steps of bayesian optimization you want to perform. The more steps the more likely to find a good maximum you are.
# optimization_round = 10 # Simply, 10 models with different parameter will be tested. And the best one will be returned.

# #init_points: How many steps of random exploration you want to perform. Random exploration can help by diversifying the exploration space.
# init_points = 2

# random_seed = 2010

# best_params= []
# for cv_estimator in cv_estimators:
#     opt_params = lgb_best_params(features, target,params_range, init_points=init_points, optimization_round=optimization_round, n_folds=fold, random_seed=random_seed, cv_estimators=cv_estimator)
#     opt_params['params']['iteration'] = cv_estimator
#     opt_params['params']['fold'] = fold
#     opt_params['params']['rmse'] = opt_params['target']
#     best_params.append(opt_params['params'])

# df_params = pd.DataFrame(best_params).reset_index()
# df_params = df_params[['iteration','fold','num_leaves','learning_rate','bagging_fraction',
#  'feature_fraction',
#  'lambda_l1',
#  'lambda_l2',
#  'max_depth',
#  'min_child_weight',
#  'min_split_gain',
#  'rmse']]
# df_params.to_csv('best_params.csv')
